refactor: flatten evaluator inheirtance
