+++ tests/unit/executors/evaluators/rank/test_recall.py
@@ -1,6 +1,6 @@
 import pytest
 
-from jina.executors.evaluators.ranking.recall import RecallEvaluator
+from jina.executors.evaluators.rank.recall import RecallEvaluator
 
 
 @pytest.mark.parametrize(
